llm_info_aggregator/requirements.txt
streamlit>=1.33.0
requests>=2.31.0
beautifulsoup4>=4.12.3
pandas>=2.0.0
python-docx>=1.1.0
PyPDF2>=3.0.1
reportlab>=4.0.0
openai>=1.30.0
python-dateutil>=2.9.0.post0
